YLW-3184: Moved actions in to actions folder
